refactor: move package to `domain`
