+++ src/main/java/lotto/domain/LottoProblem.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.util.BaseProblem;
 import lotto.util.Global;
