+++ src/test/java/lotto/domain/WinningStatsTest.java
@@ -1,12 +1,13 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
 import java.util.List;
+
 import org.junit.jupiter.api.Test;
 
-public class WinningStatsTest {
+class WinningStatsTest {
     @Test
     void testToString() {
